feat(ui): UI enhancements
closes gravitee-io/issues#2357
