+++ gravitee-am-ui/src/app/app.module.ts
@@ -65,6 +65,7 @@ import { ClientComponent } from './domain/clients/client/client.component';
 import { ClientCreationComponent } from './clients/creation/client-creation.component';
 import { ClientSettingsComponent } from './domain/clients/client/settings/settings.component';
 import { ClientIdPComponent } from './domain/clients/client/idp/idp.component';
+import { ClientOAuth2Component } from './domain/clients/client/oauth2/oauth2.component';
 import { ClientOIDCComponent, CreateClaimComponent } from './domain/clients/client/oidc/oidc.component';
 import { ClientEmailsComponent } from './domain/clients/client/emails/emails.component';
 import { ClientEmailComponent } from './domain/clients/client/emails/email/email.component';
@@ -238,6 +239,7 @@ import { PoliciesResolver } from "./resolvers/policies.resolver";
     ClientComponent,
     ClientCreationComponent,
     ClientSettingsComponent,
+    ClientOAuth2Component,
     ClientOIDCComponent,
     ClientIdPComponent,
     ClientEmailsComponent,
