+++ gravitee-am-ui/src/app/clients/creation/client-creation.component.scss
@@ -1,10 +1,17 @@
 .new-client-container {
   padding: 20px 70px 20px 50px;
+  a.back-link {
+    color: #006CAB;
+    text-decoration: none;
+    &:hover {
+      color: #2196f3;
+      text-decoration: underline;
+    }
+  }
   h1 {
     font-weight: 400;
   }
   .new-client-content {
-    padding: 45px 0 20px;
     form {
       padding: 30px 20px 20px 20px;
       background: #fafafa;
