+++ gravitee-am-ui/src/app/domain/clients/client/client.component.html
@@ -16,6 +16,7 @@
 
 -->
 <div class="client-container">
+  <a [routerLink]="['..']" class="back-link"><small><< Back to clients</small></a>
   <h1>{{client.clientName}}</h1>
   <nav mat-tab-nav-bar>
     <a mat-tab-link
