+++ gravitee-am-ui/src/app/domain/clients/client/client.component.ts
@@ -28,6 +28,7 @@ export class ClientComponent implements OnInit {
   navLinks: any = [
     {'href': 'settings' , 'label': 'Settings'},
     {'href': 'idp' , 'label': 'Identity Providers'},
+    {'href': 'oauth2' , 'label': 'OAuth 2.0'},
     {'href': 'oidc' , 'label': 'OpenID Connect'},
     {'href': 'forms' , 'label': 'Forms'},
     {'href': 'emails' , 'label': 'Emails'},
