+++ gravitee-am-ui/src/app/domain/clients/client/oidc/claims/add-claim.component.html
@@ -15,20 +15,16 @@
     limitations under the License.
 
 -->
-<h5 matDialogTitle>
-  Add ID Token claim
-</h5>
-<mat-dialog-content>
-  <form #claimForm="ngForm">
-    <mat-form-field>
-      <input matInput name="key" placeholder="Claim name" required autofocus ngModel/>
-    </mat-form-field>
-    <mat-form-field>
-      <input matInput name="value" placeholder="Claim value" required ngModel/>
-    </mat-form-field>
-  </form>
-</mat-dialog-content>
-<mat-dialog-actions>
-  <button type="button" mat-raised-button (click)="dialogRef.close(claimForm.value)" color="primary" [disabled]="claimForm.invalid">Save</button>
-  <button type="button" mat-button (click)="dialogRef.close()" class="dialog-cancel" style="margin-left: 20px;">Cancel</button>
-</mat-dialog-actions>
+<form fxLayout="row" #claimForm="ngForm" style="padding: 20px; border: 1px solid black; margin-top: 10px;">
+  <mat-form-field style="margin-right: 20px;">
+    <input matInput name="key" placeholder="Claim name" required [(ngModel)]="claim.key"/>
+    <mat-hint>The claim's name that will be displayed in the ID Token.</mat-hint>
+  </mat-form-field>
+  <mat-form-field style="margin-right: 20px;">
+    <input matInput name="value" placeholder="Claim value" required [(ngModel)]="claim.value"/>
+    <mat-hint>The claim's value contained in the user profile information.</mat-hint>
+  </mat-form-field>
+  <div>
+    <button style="margin-top: 10px;" mat-stroked-button [disabled]="claimForm.invalid || claimForm.pristine" (click)="addClaim()">+ ADD</button>
+  </div>
+</form>
