+++ gravitee-am-ui/src/app/domain/clients/client/settings/settings.component.scss
@@ -59,39 +59,11 @@ mat-icon.delete-chip {
   color: rgba(0, 0, 0, 0.38);
 }
 
-.type-section {
-  width: 50%;
-}
-
-.type-selection {
-  display: inline-flex;
-  width: 100%;
-}
-
 .client-oauth-zone {
   padding: 10px;
   background-color: #f1f1f1;
   margin-top: 20px;
   h4 {
     margin: 5px 0;
-  }
-}
-
-.client-scopes {
-  a {
-    cursor: pointer;
-    font-size: 75%;
-    color: #2196f3;
-  }
-
-  ngx-datatable {
-    background: transparent;
-    box-shadow: none;
-  }
-
-  .client-scope-actions {
-    button {
-      top: -7px;
-    }
   }
 }
