+++ gravitee-am-ui/src/app/domain/settings/account/account.component.scss
@@ -4,7 +4,6 @@
     font-weight: 400;
   }
   .domain-account-content {
-    padding: 50px 0 20px;
     .domain-account-description {
       margin-left: 20px;
       h3 {
