+++ gravitee-am-ui/src/app/domain/settings/certificates/certificate/certificate.component.scss
@@ -1,5 +1,13 @@
 .certificate-container {
   padding: 20px 70px 20px 50px;
+  a.back-link {
+    color: #006CAB;
+    text-decoration: none;
+    &:hover {
+      color: #2196f3;
+      text-decoration: underline;
+    }
+  }
   h1 {
     font-weight: 400;
   }
@@ -7,32 +15,42 @@
     margin: 0;
   }
   .certificate-content {
-    padding: 50px 0 20px;
-    .certificate-description {
-      margin-left: 20px;
-      h3 {
-        font-weight: 400;
+    form {
+      padding: 20px;
+      background: #fafafa;
+      border-radius: 2px;
+      border: 1px solid rgb(226, 229, 231);
+      .certificate-actions {
+        margin-top: 20px;
       }
-      .certificate-description-content {
-        border-top: 1px solid rgb(226, 229, 231);
-        p {
-          margin-top: 20px;
-          font-size: 14px;
-        }
-        small {
-          font-size: 12px;
-          color: rgba(0, 0, 0, 0.38);
-        }
+    }
+    .certificate-delete-zone {
+      color: #f44336;
+      font-size: 75%;
+      margin-top: 20px;
+      padding: 20px;
+      border-radius: 2px;
+      border: 1px solid #e2e5e7;
+      .certificate-delete-button {
+        margin-top: 20px;
       }
     }
   }
-  form {
-    padding: 20px;
-    background: #fafafa;
-    border-radius: 2px;
-    border: 1px solid rgb(226, 229, 231);
-    .certificate-actions {
-      margin-top: 20px;
+  .certificate-description {
+    margin-left: 20px;
+    h3 {
+      font-weight: 400;
+    }
+    .certificate-description-content {
+      border-top: 1px solid rgb(226, 229, 231);
+      p {
+        margin-top: 20px;
+        font-size: 14px;
+      }
+      small {
+        font-size: 12px;
+        color: rgba(0, 0, 0, 0.38);
+      }
     }
   }
 }
