+++ gravitee-am-ui/src/app/domain/settings/certificates/certificate/certificate.component.ts
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from "@angular/router";
+import { ActivatedRoute, Router } from "@angular/router";
 import { BreadcrumbService } from "../../../../../libraries/ng2-breadcrumb/components/breadcrumbService";
 import { PlatformService } from "../../../../services/platform.service";
 import { CertificateService}  from "../../../../services/certificate.service";
 import { SnackbarService } from "../../../../services/snackbar.service";
+import { DialogService } from "../../../../services/dialog.service";
 
 @Component({
   selector: 'app-certificate',
@@ -34,8 +35,13 @@ export class CertificateComponent implements OnInit {
   certificateConfiguration: any;
   updateCertificateConfiguration: any;
 
-  constructor(private route: ActivatedRoute, private breadcrumbService: BreadcrumbService, private platformService: PlatformService,
-              private certificateService: CertificateService, private snackbarService: SnackbarService) { }
+  constructor(private route: ActivatedRoute,
+              private breadcrumbService: BreadcrumbService,
+              private platformService: PlatformService,
+              private certificateService: CertificateService,
+              private snackbarService: SnackbarService,
+              private router: Router,
+              private dialogService: DialogService) { }
 
   ngOnInit() {
     this.domainId = this.route.snapshot.parent.parent.params['domainId'];
@@ -52,6 +58,20 @@ export class CertificateComponent implements OnInit {
       this.breadcrumbService.addFriendlyNameForRouteRegex('/domains/'+this.domainId+'/settings/certificates/'+this.certificate.id+'$', this.certificate.name);
       this.snackbarService.open("Certificate updated");
     })
+  }
+
+  delete(event) {
+    event.preventDefault();
+    this.dialogService
+      .confirm('Delete Certificate', 'Are you sure you want to delete this certificate ?')
+      .subscribe(res => {
+        if (res) {
+          this.certificateService.delete(this.domainId, this.certificate.id).subscribe(() => {
+            this.snackbarService.open("Certificate deleted");
+            this.router.navigate(['/domains', this.domainId, 'settings', 'certificates']);
+          });
+        }
+      });
   }
 
   enableCertificateUpdate(configurationWrapper) {
