+++ gravitee-am-ui/src/app/domain/settings/certificates/creation/certificate-creation.component.html
@@ -16,6 +16,7 @@
 
 -->
 <div class="new-certificate-container" fxLayout="column">
+  <a [routerLink]="['..']" class="back-link"><small><< Back to certificates</small></a>
   <h1>New certificate</h1>
   <div fxLayout="row">
     <mat-horizontal-stepper linear fxFlex="70" #stepper="matHorizontalStepper">
