+++ gravitee-am-ui/src/app/domain/settings/certificates/creation/certificate-creation.component.scss
@@ -1,5 +1,13 @@
 .new-certificate-container {
   padding: 20px 70px 0px 50px;
+  a.back-link {
+    color: #006CAB;
+    text-decoration: none;
+    &:hover {
+      color: #2196f3;
+      text-decoration: underline;
+    }
+  }
   h1 {
     font-weight: 400;
   }
