+++ gravitee-am-ui/src/app/domain/settings/general/general.component.scss
@@ -4,7 +4,6 @@
     font-weight: 400;
   }
   .domain-general-content {
-    padding: 50px 0 20px;
     form {
       padding: 20px;
       background: #fafafa;
