+++ gravitee-am-ui/src/app/domain/settings/groups/creation/group-creation.component.scss
@@ -1,10 +1,17 @@
 .new-group-container {
   padding: 20px 70px 20px 50px;
+  a.back-link {
+    color: #006CAB;
+    text-decoration: none;
+    &:hover {
+      color: #2196f3;
+      text-decoration: underline;
+    }
+  }
   h1 {
     font-weight: 400;
   }
   .new-group-content {
-    padding: 50px 0 20px;
     form {
       padding: 20px;
       background: #fafafa;
