+++ gravitee-am-ui/src/app/domain/settings/providers/creation/provider-creation.component.html
@@ -16,6 +16,7 @@
 
 -->
 <div class="new-provider-container" fxLayout="column">
+  <a [routerLink]="['..']" class="back-link"><small><< Back to identity providers</small></a>
   <h1>New provider</h1>
   <div fxLayout="row">
     <mat-horizontal-stepper linear fxFlex="70" #stepper="matHorizontalStepper">
