+++ gravitee-am-ui/src/app/domain/settings/providers/provider/mappers/mappers.component.html
@@ -16,59 +16,69 @@
 
 -->
 <div class="provider-mappers-container">
-  <div fxLayout="column">
-    <small>
-      Attributes mapping let you bind user profile attributes with your identity provider attributes.
-      These new attributes specified in this mapping can be retrieved via the UserInfo Endpoint or be used with ID Token claims.
-      This mapping occurs during user authentication.
-    </small>
+  <div class="provider-mappers-content" fxLayout="column" fxFlex="70">
     <small style="margin-top: 10px;" *ngIf="!isEmpty" class="note"><i>Double click to edit and press enter to save changes</i></small>
-  </div>
-  <ngx-datatable *ngIf="!isEmpty" [columnMode]="'flex'" class="material" [headerHeight]="40" [footerHeight]="40" [rowHeight]="55" [rows]='mappers'>
-    <ngx-datatable-column name="User profile attributes" [flexGrow]="1">
-      <ng-template let-row="row" let-rowIndex="rowIndex" ngx-datatable-cell-template>
+    <ngx-datatable *ngIf="!isEmpty" [columnMode]="'flex'" class="material" [headerHeight]="40" [footerHeight]="40" [rowHeight]="55" [rows]='mappers'>
+      <ngx-datatable-column name="User profile attributes" [flexGrow]="1">
+        <ng-template let-row="row" let-rowIndex="rowIndex" ngx-datatable-cell-template>
         <span (dblclick)="editing[rowIndex + '-key'] = true" *ngIf="!editing[rowIndex + '-key']">
           {{row.key}}
         </span>
-        <mat-form-field *ngIf="editing[rowIndex + '-key']" class="datatable-input">
-          <input matInput type="text" required autofocus placeholder="User attribute"
-            (keyup.enter)="updateMapper($event, 'key', rowIndex)"
-            (blur)="editing[rowIndex + '-key'] = false"
-            [value]="row.key" />
-        </mat-form-field>
-      </ng-template>
-    </ngx-datatable-column>
-    <ngx-datatable-column name="Raw attributes" [flexGrow]="1">
-      <ng-template let-row="row" let-rowIndex="rowIndex" ngx-datatable-cell-template>
+          <mat-form-field *ngIf="editing[rowIndex + '-key']" class="datatable-input">
+            <input matInput type="text" required autofocus placeholder="User attribute"
+                   (keyup.enter)="updateMapper($event, 'key', rowIndex)"
+                   (blur)="editing[rowIndex + '-key'] = false"
+                   [value]="row.key" />
+          </mat-form-field>
+        </ng-template>
+      </ngx-datatable-column>
+      <ngx-datatable-column name="Raw attributes" [flexGrow]="1">
+        <ng-template let-row="row" let-rowIndex="rowIndex" ngx-datatable-cell-template>
         <span (dblclick)="editing[rowIndex + '-value'] = true" *ngIf="!editing[rowIndex + '-value']">
           {{row.value}}
         </span>
-        <mat-form-field *ngIf="editing[rowIndex + '-value']" class="datatable-input">
-          <input matInput type="text" required autofocus placeholder="Raw attribute"
-            (keyup.enter)="updateMapper($event, 'value', rowIndex)"
-            (blur)="editing[rowIndex + '-value'] = false"
-            [value]="row.value" />
-        </mat-form-field>
-      </ng-template>
-    </ngx-datatable-column>
-    <ngx-datatable-column name="Actions" [flexGrow]="1">
-      <ng-template let-row="row" ngx-datatable-cell-template>
-        <div fxLayout="row" class="provider-mappers-actions">
-          <button mat-icon-button (click)="delete(row.key, $event)"><mat-icon>delete</mat-icon></button>
-        </div>
-      </ng-template>
-    </ngx-datatable-column>
-  </ngx-datatable>
+          <mat-form-field *ngIf="editing[rowIndex + '-value']" class="datatable-input">
+            <input matInput type="text" required autofocus placeholder="Raw attribute"
+                   (keyup.enter)="updateMapper($event, 'value', rowIndex)"
+                   (blur)="editing[rowIndex + '-value'] = false"
+                   [value]="row.value" />
+          </mat-form-field>
+        </ng-template>
+      </ngx-datatable-column>
+      <ngx-datatable-column name="Actions" [flexGrow]="1">
+        <ng-template let-row="row" ngx-datatable-cell-template>
+          <div fxLayout="row" class="provider-mappers-actions">
+            <button mat-icon-button (click)="delete(row.key, $event)"><mat-icon>delete</mat-icon></button>
+          </div>
+        </ng-template>
+      </ngx-datatable-column>
+    </ngx-datatable>
 
-  <app-emptystate *ngIf="isEmpty"
-                  [message]="'Identity provider mappers will appear here'"
-                  [subMessage]="'Start creating a mapper'"
-                  [icon]="'compare_arrows'"></app-emptystate>
+    <app-emptystate *ngIf="isEmpty"
+                    [message]="'Identity provider mappers will appear here'"
+                    [subMessage]="'Start creating a mapper'"
+                    [icon]="'compare_arrows'"></app-emptystate>
 
-  <div [ngClass]="{'add-button': !isEmpty, 'add-button-center': isEmpty}">
-    <button mat-fab color="accent" (click)="add()">
-      <mat-icon>add</mat-icon>
-    </button>
+    <div [ngClass]="{'add-button': !isEmpty, 'add-button-center': isEmpty}">
+      <button mat-fab color="accent" (click)="add()">
+        <mat-icon>add</mat-icon>
+      </button>
+    </div>
+  </div>
+
+  <div class="provider-mappers-description" fxFlex>
+    <h3>User mappers</h3>
+    <div class="provider-mappers-description-content">
+      <p>
+        Attributes mapping let you bind user profile attributes with your identity provider attributes.
+      </p>
+      <p>
+        These new attributes specified in this mapping can be retrieved via the UserInfo Endpoint or be used with ID Token claims.
+      </p>
+      <p>
+        This mapping occurs during user authentication.
+      </p>
+    </div>
   </div>
 </div>
