+++ gravitee-am-ui/src/app/domain/settings/providers/provider/mappers/mappers.component.scss
@@ -2,6 +2,24 @@
   small.note {
     color: grey;
   }
+
+  .provider-mappers-description {
+    margin-left: 20px;
+    h3 {
+      font-weight: 400;
+    }
+    .provider-mappers-description-content {
+      border-top: 1px solid rgb(226, 229, 231);
+      p {
+        margin-top: 20px;
+        font-size: 14px;
+      }
+      small {
+        font-size: 12px;
+        color: rgba(0, 0, 0, 0.38);
+      }
+    }
+  }
 }
 
 .ngx-datatable.material {
