+++ gravitee-am-ui/src/app/domain/settings/providers/provider/provider.component.html
@@ -16,6 +16,7 @@
 
 -->
 <div *ngIf="provider" class="provider-container">
+  <a [routerLink]="['..']" class="back-link"><small><< Back to identity providers</small></a>
   <h1>{{provider.name}}</h1>
   <nav mat-tab-nav-bar>
     <a mat-tab-link
