+++ gravitee-am-ui/src/app/domain/settings/providers/provider/provider.component.scss
@@ -1,5 +1,13 @@
 .provider-container {
   padding: 20px 70px 20px 50px;
+  a.back-link {
+    color: #006CAB;
+    text-decoration: none;
+    &:hover {
+      color: #2196f3;
+      text-decoration: underline;
+    }
+  }
   h1 {
     font-weight: 400;
   }
