+++ gravitee-am-ui/src/app/domain/settings/providers/provider/roles/roles.component.scss
@@ -1,5 +1,19 @@
-.provider-roles-example {
-  margin-top: 10px;
+.provider-roles-description {
+  margin-left: 20px;
+  h3 {
+    font-weight: 400;
+  }
+  .provider-roles-description-content {
+    border-top: 1px solid rgb(226, 229, 231);
+    p {
+      margin-top: 20px;
+      font-size: 14px;
+    }
+    small {
+      font-size: 12px;
+      color: rgba(0, 0, 0, 0.38);
+    }
+  }
 }
 
 .add-button {
