+++ gravitee-am-ui/src/app/domain/settings/roles/creation/role-creation.component.scss
@@ -1,10 +1,17 @@
 .new-role-container {
   padding: 20px 70px 20px 50px;
+  a.back-link {
+    color: #006CAB;
+    text-decoration: none;
+    &:hover {
+      color: #2196f3;
+      text-decoration: underline;
+    }
+  }
   h1 {
     font-weight: 400;
   }
   .new-role-content {
-    padding: 50px 0 20px;
     form {
       padding: 20px;
       background: #fafafa;
