+++ gravitee-am-ui/src/app/domain/settings/roles/role/role.component.scss
@@ -1,10 +1,17 @@
 .role-container {
   padding: 20px 70px 20px 50px;
+  a.back-link {
+    color: #006CAB;
+    text-decoration: none;
+    &:hover {
+      color: #2196f3;
+      text-decoration: underline;
+    }
+  }
   h1 {
     font-weight: 400;
   }
   .role-content {
-    padding: 50px 0 20px;
     form {
       padding: 20px;
       background: #fafafa;
@@ -14,21 +21,32 @@
         margin-top: 20px;
       }
     }
-    .role-description {
-      margin-left: 20px;
-      h3 {
-        font-weight: 400;
+    .role-delete-zone {
+      color: #f44336;
+      font-size: 75%;
+      margin-top: 20px;
+      padding: 20px;
+      border-radius: 2px;
+      border: 1px solid #e2e5e7;
+      .role-delete-button {
+        margin-top: 20px;
       }
-      .role-description-content {
-        border-top: 1px solid rgb(226, 229, 231);
-        p {
-          margin-top: 20px;
-          font-size: 14px;
-        }
-        small {
-          font-size: 12px;
-          color: rgba(0, 0, 0, 0.38);
-        }
+    }
+  }
+  .role-description {
+    margin-left: 20px;
+    h3 {
+      font-weight: 400;
+    }
+    .role-description-content {
+      border-top: 1px solid rgb(226, 229, 231);
+      p {
+        margin-top: 20px;
+        font-size: 14px;
+      }
+      small {
+        font-size: 12px;
+        color: rgba(0, 0, 0, 0.38);
       }
     }
   }
