+++ gravitee-am-ui/src/app/domain/settings/users/user/user.component.html
@@ -16,6 +16,7 @@
 
 -->
 <div class="user-container">
+  <a [routerLink]="['..']" class="back-link"><small><< Back to users</small></a>
   <h1>{{user.username}}</h1>
   <nav mat-tab-nav-bar>
     <a mat-tab-link
